+++ src/Service/S3/src/S3Client.php
@@ -3,6 +3,7 @@
 namespace AsyncAws\S3;
 
 use AsyncAws\Core\AbstractApi;
+use AsyncAws\Core\Configuration;
 use AsyncAws\Core\RequestContext;
 use AsyncAws\Core\Result;
 use AsyncAws\S3\Enum\BucketCannedACL;
@@ -939,8 +940,10 @@ class S3Client extends AbstractApi
     protected function getSignerFactories(): array
     {
         return [
-            's3v4' => static function (string $service, string $region) {
-                return new SignerV4ForS3($service, $region);
+            's3v4' => function (string $service, string $region) {
+                $configuration = $this->getConfiguration();
+
+                return new SignerV4ForS3($service, $region, !Configuration::optionExists('sendChunkedBody') || \filter_var($configuration->get('sendChunkedBody'), \FILTER_VALIDATE_BOOLEAN));
             },
         ] + parent::getSignerFactories();
     }